Reverting to single distance_to_zone.py file.
