--- baseballcv/functions/utils/distance_to_zone/__init__.py
@@ -1,3 +1,0 @@
-from .base import DistanceToZone
-
-__all__ = ['DistanceToZone']
